requirements.txt
chardet==2.3.0
diff-match-patch==20121119
Django==1.9.4
django-allauth==0.24.1
django-filter==0.12.0
django-filters==0.2.1
django-import-export==0.4.5
django-localflavor==1.2
django-rest-auth==0.7.0
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
Markdown==2.6.5
oauthlib==1.0.3
Pillow==3.1.1
psycopg2==2.6.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
tablib==0.11.2
